+++ src/clojure/nrepl/threads.clj
@@ -1,0 +1,4 @@
+(ns nrepl.threads)
+
+(def ^{:dynamic true
+       :tag java.util.concurrent.ThreadFactory} *platform-thread-factory* nil)